Added JSON model for floor "quad"
Fades in and out with rotation, which I'm pretty sure should not be
happening, but at least it's there. Set a dimension for the camera so it
does not zoom in and out when choose a new renderable. Renderables still
in center position on top of each other. Will eventually take out the
check boxes for static scene.